Mudando a estrutura do codigo
